Split sprite into renderer, sprite, & spritesystem
